requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==5.0.0
boto3==1.42.78
botocore==1.42.78
certifi==2026.2.25
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
Deprecated==1.3.1
ecdsa==0.19.2
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
jmespath==1.1.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pyasn1==0.6.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wrapt==2.1.2
